Initial code upload on ImsStack
This change is to upload initial code for ImsStack modules.
ImsStack modules consist of one apk and one shared library.
    - ImsStack(apk) / libimsstack

BYPASS_INCLUSIVE_LANGUAGE_REASON=details below
The following words are from the standard specification to make the code
more clear and these words(“dummy-rfc5367”/”ipsec-man”) should be
included in the SIP message to meet the standard and carrier's
requirement.
    “dummy-rfc5367”(IETF)
    “ipsec-man”(3GPP/IETF)

Test: manual (verify that ImsStack runs on cuttlefish)
Change-Id: I149ae2ed87c4d08e3c11efe6b726c4ad26aba697