Change class/method naming (code-style) - engine/sipcore
This changes the class/method naming for code style in
engine/sipcore/interface component.
    - use pascal case (i.e. SIP -> Sip)

BYPASS_INCLUSIVE_LANGUAGE_REASON=details below
The following word is from the standard specification to make the code more clear and this word("ipsec-man”) should be included in the SIP message to meet the standard and carrier's requirement.
    “ipsec-man”(3GPP/IETF)

Test: m (build)
Change-Id: I343484b8223b1f52437bd9ca0f1225ba5a553390