(Mtc) Fix not to destroy the media profiles on call establishment
MTC destroyed the media profiles which are not used in confirmed state
when the call is established.
As redeifining DestroyProfile() to use when the specific dialog is terminated
with BYE request within a call, fix not to destroy the unused profiles
in other cases.
And change the calling time for creation of media session and profile in MT call case.
It is called after the call type is updated based on incoming INVITE request.

Test: mm build and manual test with the media module (ag/18792849)
Change-Id: I4d7d8d5aa2f6647c67bfcfd1b1cab698c84a44fe