Fix abnormal behavior on session refresh helper
Since there is no explicit include statement in SessionRefreshHelper, it
seems that the memory address of the pointer variable(SipConfigV*) was
changed when the INVITE request was sent. To avoid this situation, the
include statement for SipConfigV is added to the Service class that
provides the SipConfigV to the SessionRefreshHelper.

Test: manual (verify ims-call on device)
Change-Id: Ia37e72514ddaacfc5db0413fd059725af4ed85d6