Refactor platform interface for unit test
This commit is to refactor the platform interface to write the unit test
code easily by mocking the platform interfaces.
Most singleton objects are collected and managed by the PlatformContext
that is newly added. This change also adds the basic testable classes
that can be used by the unit tests under the tests directory.

Bug: 224463364
Test: atest ImsStackNativeTests, atest ImsStackNativePlatformTests
Test: manual (verify ims-reg/call on target device)
Change-Id: Ic53877c4a53a573ca41a13e25392cff3fec45c39