Update JNI method naming/direct native method call
This commit is to update the JNI method naming and the direct native
method call in the Java logic. And, it also contains the newly
introduced method to control and manage the native logic directly. The
setConfiguration method is kept temporarily and will be removed after
adding the native control(start/stop enablers) logic.

Bug: 241188001
Test: atest ImsStackJavaTests
Test: manual (verify ims-reg/call on target device)
Change-Id: I1c9c8399a9d419ffbe74a72310b4b8eec0cb9166