Add a config for controlling "text" feature-tag
This commit is to add a configuration and its related logic for
controlling "text" feature-tag. The "text" feature-tag is added as a
part of Stream property and it's configurable using ISipConfigV
interface whether it needs to be added as a feature-tag or not.
Like a "video" feature-tag, if this feature-tag option is set or clear,
then the caller capability in IRegContact should be recalculated to
apply the feature-tag options on the runtime.

Bug: 242010794
Test: atest ImsStackNativeTests
Test: manual (verify ims-reg/call on target device)
Change-Id: I74ccb9223a72fa67d0719993bf32d941972f65ac