Propagate carrier-config when IMS service is already running
After ImsStack is started by the SIM state change, it didn't proceed
the IMS registration because the VoLTE is not available at this point.
After a while, the carrier-config is changed, but it's not passed to the
internal components (ConfigAgent, native layer). So, IMS registration
over this SIM slot was not started even if the VoLTE is available now.
This commit is to propagate the carrier-config change to the internal
components (ConfigAgent, native layer) if the carrier-config is changed
while the IMS service is already running. This will allow the IMS
service to proceed with the IMS registration process.

Bug: 257157485
Test: manual (verify ims-reg in single or dual mode)
Test: atest ImsStackJavaTests
Change-Id: I16284cd5c9e36bfc6f671ba282753a168aaaf538