(Mtc) Unblock MO call when Ongoing Call Upgrade exists
This commit is to fix a bug. Currently, MO call is blocked if
another call exists and its state is UpdatingState. But,
this makes adding call fail in case the network send offerless
re-INVITE just after HOLD operation/transaction is completed.
So, changing UE behavior not to block an MO call in the case.

Bug: b/259377313
Test: atest passed, simulated using DISS
Change-Id: I19aff12789838b63d020592e790ebeece53d7b96