(Mtc) bug fix - video direction depends on audio direction
During the VZW video direction test, one issue was observed.
issue : if audio direction is sendrecv and video direction is inactive
and the UE receives re-INVITE with audio=sendonly and video=inactive,
the UE answers with audio recvonly and video recvonly.
The expectation is audio recvonly and video inactive.
Audio direction must affect video direction only in case of
receiving offereless re-INVITE so the logic is separated.

Bug: b/242151253
Test: simluation verified. atest passed
Change-Id: Ic40b657908dee20f2c5497f016c336c21d61bb06