sip: Fix txn testcases failed when ran individually
For code coverage purpose mock start timer api should return once false and once true.
Static varible count was used for this purpose when count is 0 return false and when true return true.
But was not reset before starting each testcase. So if ran individually failing if ran together passing.
So reset in setup api and reordered the testcase, so that passes both running individually and together as well.

Test: atest ImsStackNativeProtocolSipTests
Bug: 261372516
Change-Id: Iba757e60768e5224055aedaf40c248283d60ffe6