(Ut) Don't remove SscServieImpl in clear() of ImsUtImpl
This is a CL not to remove SscServiceImpl when clear() of ImsUtImpl
invoked. clear() function should just handle clearing operation and
close() function should have a role of removing SscServiceImpl.
The problem is that if SscServiceImpl is removed in clear(),
ServiceServiceStateListener set by ImsFeatureManager also removed. The
listener is set only when ImsUtImpl is created, and it should be removed
when ImsUtImpl is destoryed.

Bug: 271355203
Test: atest ImsStackJavaTests
Change-Id: Ib6941336a636cef09cc5771ed9b8e2263fae9a40