(Mtc) Update MtcPreconditionManager series to support add media handling
This commit enhances MtcPreconditoinManager to support add and remove
additional media type and their QoS during upgrade and downgrade.
To do this, it needs to check the current call type's media types so
SdpPreconditionHelper#GetMediaTypesByBdp is replaced by
GetMediaTypesFromCallType. And It intializes the QoS status to IDLE for
unused media types.
Note: the caller (UpdatingState) should invoke OnCallModified() for
every cases of changing the CallType.

Bug: 274840426
Test: atest ImsStackNativeEnablerMtcTests
Test: verified using temp code for add video call flow

Change-Id: I7927b30a01212bec63b52b762675517edac597a3