test: use the SIM carrier configuration if carrier-ids match
Currently, the carrier configuration is provided by changing the
settings required for testing based on the default carrier
configuration.
We will provide one more option to manage and control the carrier
configuration more easily.
If the carrier-id of SIM currently inserted into the device matches the
carrier-id of the test, this change uses the SIM carrier configuration
instead of the default carrier configuration as the default test
configuration. Therefore, if we need to test according to the
configuration of a specific carrier, just set the carrier-id value that
matches the SIM to be used in the device to test carrier-id.

Bug: 308693157
Test: atest com.android.imsstack.its.tests.BootupTest
Change-Id: I5e139728bc1f74fa776e31694c6a77550e0cb0cc