(Mtc) Fix unexpected conference event handling bug
This is a fix for an issue where the 'disconnected' event was not
handled properly when the user entity for the last remaining user in the
full state Conference event package removed(i.e., the state was not set
to 'deleted'). Currently, if there are no user entities, it is
recognized as an initial conference event package, and the user list is
not updated. This has been modified to additionally verify if it is the
first Conference event package by checking the version.

Test: Tiss simulation using cl/650796408
Test: atest ImsStackNativeTests
Bug: 352170872
Flag: NONE pre-production
Change-Id: I3e17e59371c1529c58875f9006c954038a6cf0a3