(Mtc) Fix bug : non SIP protocol type Reason header is used
When a CANCEL message is received with multiple Reason headers having
different protocols, the validation is performed using the first
protocol. In this case, if the Reason header with the SIP protocol is
not the topmost Reason header, the IMS Stack fails to find the
appropriate Reason and malfunctions. This CL fixess that issue.

Test: simulation using TISS cl/651598382
Bug: 352188649
Flag: NONE non-production
Change-Id: I50ac2d97571a194e87c21737119cd72c950f6cc5