test: Add setting verdict condition and notifying the result
This change adds an implementation that uses ControlConnection to send
test verdict conditions to the server and receive the corresponding
results. Users can set the ruleset using ClientMessage, and by
implementing the ServerFailureHandler class and passing it to
ControlConnection, they can receive the server's results if the test is
failed.

Test: ImsStackTest using test code.
Flag: NONE pre-production
Bug: 369467734
Change-Id: I3f6d11191b30db9d572d15146b87daf42cb7d021