(Mtc) Use a `TestTimerService` instead of directly setting a `MockIOsFactory`
When makes the `TimerService`, a `ImsMutex` is created by a `IOsFactory`.
If the `MockIOsFactory` isn't set for creating the `ImsMutex`, that
makes problem like this.
The reason why issue was not occurred before is the `TimerService` is
already mede by another test.

Test: atest ImsStackNativeEnablerMtcTests, atest ImsStackNativeTests,
each test that makes peoblem
Bug: b/372122530
Flag: NONE pre-production

Change-Id: I2dac1c7152eb61e56c87691e5d932b4f62cd2c64