(Mtc) Change some files importing format
Based on Android.bp file definition in the AP IMS, dont' need to add
additional directory path to use files in the another area if files in
the declared path location.
please see details about directory structure of the AP IMS in the
document in the comment.

Bug: b/350846576
Test: atest ImsStackNativeEnablerMtcTests, atest ImsStackNativeEnablerMtsTests
Flag: NONE pre-production
Change-Id: I85aafa4093cbd58e9aa88bec8ce65b3520186421