(Mtc) Change emergency variable in CallInfo to enum type - Native
This change proposes converting the emergency variable in CallInfo from
a boolean type to an enumeration.
This will allow the emergency variable to convey not only the service
type information from Telephony, but also the emergency routing
information.
Instead of activating based on the existing 'bEmergency' variable being
true, the functionalities will be triggered only for calls received from
Telephony with a service type of emergency and a routing type of unknown
or emergency.
The behavior of processing normal routing calls requested by Telephony
as emergency calls will be modified later for each scenario.

Test: manual test with TISS
Test: atest ImsStackJavaTests
Test: atest ImsStackNativeTests
Bug: 365713214
Flag: NONE pre-production
Change-Id: Ice8a9072a64c3ed889850e71cdac4ea37327bdcf