(Mtc) Notifies the Telephony with currently attached RAT during calling
Notifying the Telephony with a ImsCallProfile which has a specific
call extra(EXTRA_CALL_NETWORK_TYPE).
Each MO/MT has a first callback which carries that profile. Before that
callback, another callback to update that call extra value is not
allowed.
Please see the design document in the bug for details.

Bug: b/378010483
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests, atest ImsStackNativeEnablerMtcTests
Change-Id: I0375486be0c5b9dea784addf5a8ad971237d5fd6