(sipcore/platform) Reuse only currently alive SIP sockets
Generally, the SIP transport layer reuses TCP sockets for the same SIP
server. Due to timing, if a SIP socket is closed and reused immediately,
an unnecessary transport layer error is caused by reusing the socket
that will be soon be closed.
To prevent this situation, this change is to add a condition to
determine whether to reuse by checking the current status of the socket
(closed or being closed) when reusing a SIP socket.

Bug: 377643100
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verify IMS reg/call on device)
Test: manual (verify GCF-VoNR TC_6.4)
Change-Id: Id95815edf06ecc664bcd6277c54e8816c568f814