(core) Fix a crash on CleanupOnDestroy
A crash occurred while each ServiceMethod(Session, Subscription,
Reference) was releasing resources after the engine's Service object was
destroyed during the Enabler restart process due to a crash of an
external module during a 1-to-1 call or conference call.
This change is to release the resource in the Destroy() method instead
of the CleanupOnDestroy() method to avoid accessing the Service object
that has already been released when accessing the Service object in the
ServiceMethod.

Bug: 379586139
Bug: 377588525
Bug: 377181201
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verify IMS reg/call on device)
Test: manual (verify that ImsStack crash is not observed when an
external module is crashed during VoLTE/VoWiFi call)

Change-Id: If6303f6656f05531b82e8622edcf160dd947676b