(Mtc) Implementation of cross SIM redialing
Upon receiving an error response to an INVITE message for an emergency
call, the device should check if the error response code indicates that
the carrier requires cross SIM redialing. If so, a cross SIM redialing
should be initiated to the Telephony module.
However, this request should only be made if the SIM in the other slot
is in an available state.

Test: atest ImsStackJavaTests
Test: atest ImsStackNativeTests
Bug: 364014358
Flag: NONE pre-production
Change-Id: I3bed33346ffc25052da5f4532a196c04fa25834a