Fix incorrect behavior for getDataNetworkType
TelephonyManager provides the getDataNetworkType API, but it returns the
data network type for a default data subscription instead of a specific
subscription or a specific slot. Especially, when there is no SIM,
ImsStack needs to get the network type based on the slot instead of
subscription.
So ImsStack gets the data/voice network type directly from the
ServiceState so that the internal components can get the network type
for a given subscription or slot.

Bug: 377447485
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests
Test: manual (verify IMS reg/call on device)
Test: manual (verify DS_EMC_032 No SIM, IMS Emergency Call)
Change-Id: I8875a021e5354a2e65539fc1f6b8310f406256fa