(Mtc) Prevent duplicate deletion of OperationAsyncRunner
OperationAsyncRunner could be deleted twice when the operation set in
OperationAsyncRunner deletes the runner itself, leading to an exception.
This commit prevents the duplicate deletion by checking if the
OperationAsyncRunner is already processing an operation posted by a
thread message. If it is, the deletion is deferred to the
RemoveCallBack.

Test: normal call termination
Test: Remove by owner case simulation
Test: Remove by callback (normal)
Test: atest ImsStackNativeEnablerMtcTests
Bug: 382144292
Flag: NONE pre-production
Change-Id: Ia386c115204042dc6f68f9dc944ea10a31b9496b