(Mtc) Redial with VoIP from RTT emergency rejection
This change is add the behavior to re-dial with VoIP when RTT emergency
call is rejected or timed out by network if the related carrier
configuration is set as true.
To support this, the carrier configuration
`imsemergency.silent_redial_with_voip_by_rtt_rejection_bool` is added.
The default value of the configuration is false.
T-Mobile US requires this behavior, so the configuration value is set as
true.

Test: atest ImsStackNativeEnablerMtcTests
Bug: 383229932
Flag: NONE pre-production
Change-Id: I908ff283e34682ed7580b4f34fd894102c3d3ec9