(Mtc) Handle OnConnectionFailed after OnConnectionSetupPrepared
MTC doesn't handle OnConnectionFailed after OnConnectionSetupPrepared is
invoked since OnConnectionSetupPrepared implies that the traffic is
prepared for VoLTE. But, depending on the behavior of the modem side,
there is chance to receive OnConnectionFailed even after INVITE is sent.
This adds the handling logic to terminate the call in OutgoingState in
case OnConnectionFailed is called.

Test: atest ImsStackNativeEnablerMtcTests
Test: verified by TC in the bug
Bug: 385008677
Flag: NONE pre-production
Change-Id: Ic6983d9631b306958f2880a37c621e28320d9b33