(core) Add handling for configuring user-info part of Contact header
This change is to add a new configuration for the user-info part of the
Contact header used in IMS registration and add its related handling
accordingly.
The user-info part is divided into the following three types and
processed:
 - UUID (default, time-based UUID value)
 - IMPU (copy user-info part from IMPU)
 - EMPTY (no user-info part)

The anonymous emergency call is usually set to "anonymous".

Bug: 371445406
Flag: NONE pre-production
Test: atest ImsStackNativeTests
Test: manual (verified the user-info part of Contact header based on the
configuration values)

Change-Id: I2aab04e34f7d3146028b470d0b94d765668ea4ab