Delay handling of ISIM LOADED state for reading ISIM records
Currently, the Telephony framework doesn't provide an event for the
complete loading of the ISIM application. ImsStack manages the ISIM
state based on the state of USIM application. This can lead to a state
mismatch when the framework reads ISIM records while ImsStack is
simultaneously reading those fields.
This change introduces a time-based approach for handling the ISIM
LOADED state. To ensure the framework has finished reading all ISIM
records, ImsStack wil repeatedly read the ISIM records for a certain
period. If the mandatory fields are successfully loaded, the ISIM LOADED
state will be processed at that time.

Bug: 400168659
Flag: EXEMPT bugfix
Test: atest ImsStackJavaTests
Test: manual (verified IMS registration with a valid information that is
provisioned in ISIM application)

Change-Id: I8546f208cc7d4c6fd3d83342c33d8da334f71e31