Adjust the timing of the session opening
[Problem]
The session opening is done when the SDP is forming which will make the session open unnecessary when the SDP is not answered but rejected.

[Fix]
Move the session opening routine from forming the SDP to after the negotiation to avoid unnecessary session opening.
Add failure log to check the error cases

Bug: 395029653
Bug: 408934353
Test: Verified by the voice and video call MO/MT, upgrade and downgrade, atest ImsStackNativeEnablerMediaTests
Flag: EXEMPT bugfix
Change-Id: I89948d31a4202cabab216a4f46ac9a9c1d28d70d