(reg) Be lenient when validating contact URI field in reg-info XML
This change is to allow the current parsing process to continue even if
the "uri" field of "contact" element in reg-info XML document is
malformed and leave related exception handling to the enabler.
Additionally, the Engine has been added to handle the "shortened" event
based on the carrier configuration, regardless of the URI comparison
result.

Bug: 401434310
Bug: 401431033
Flag: EXEMPT bugfix
Test: manual (verified all the elements of "contact" element are
successfully parsed except for "uri" field and passed to the enabler)

Change-Id: Ib7c24dd2a0e9d9d25e86dc7d2514c26e42fbc833