(mtc) Create native MtcCall object before opening emergency service
Previously, Native MtcCall objects for emergency calls were created
after the emergency service was opened. This change modifies the creation
timing to match that of normal calls.

This change was made to allow specific logic to run before the
`MtcCall#open` method is called on the native object, and to improve
consistency between emergency and normal call handling.

Bug: 412489169
Test: atest ImsStackJavaTests
Test: Make MO/MT call, make emergency call
Flag: NONE pre-production
Change-Id: I3a91993fe059978e291b560e91e5dcbd3d61eb3a