(SIP) Fix an issue in URI decoding failure
This change is to fix decoding failure in To, From, and Contact headers when the user-info part contained a semicolon.
Angle brackets were the delimiter for URI and parameters.
In absence of angle brackets, semicolon is used as delimiter for URI and parameters. This created issue when user-info contains semicolon.
Fixed it to evaluate URI and parameters delimiter with '@' and semicolon in absence of angle bracket.

Bug: 411247095
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verified the issued testcase with TISS)
Test: manual (verified IMS reg/call on device)
Change-Id: I71dcbd245eee5ea1428a25440bebd1a8ee5d7a90