(xml) Normalize CRLF/CR characters in XML before parsing
When parsing XML using the libxml2 library, if there is a
carriage-return (CR) character in the text node, XML parsing is
successfully done, but the value of the text node cannot be parsed
properly.
Thus, this change is to replace CRLF / CR to line-feed (LF) if CR
character is included in XML so that the value of text node can be
parsed correctly.

Bug: 415720337
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verified that the value of a text node is parsed normally
when the XML contains a text node with a CR character)

Change-Id: Ide4ba331e3610e0c4e4b81819077df63bbd354eb