(Mtc) add host part to anonymous user id
This change adjusts the ImsConferenceState.USER value passed to the upper layer (telephony).

For anonymous participants, the IMS Stack internally constructs a string in the format "anonymous + index" and sets it as the USER value. In this scenario, the upper layer displays this name directly on the UI. To present it as a "Private Number" format on the UI, the value needs to be passed as "anonymous + index + @anonymous.invalid".

Test: test anonymous case and check UI
Test: atest ImsStackJavaTests::MtcCallTest
Bug: 419741485
Flag: NONE pre-production
Change-Id: I7ffaafb7e61719beac0bccc017e35219c5ea316c