(Mtc) SessionTerminated() in IdleState
This commit implements SessionTerminated() logic in IdleState.
Currently, it's not handled so MTC keeps the IdleState but there is
possibility to be invoked this ISessionListener API if no ATTACH is
passed by the MTC Java layer for the call. In the case, MTC cannot clear
the MtcCall so it infinitely rejects all MO/MT calls.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulated the scenario with test code
Bug: 426985882
Flag: NONE pre-production
Change-Id: I4d847be781f584423b2b63b527c9928ef40feac5