(SIP) Fix incorrect syntax in P-Preferred-Identity header
The P-Preferred-Identity header should basically consist of only
name-addr or addr-spec, but the header parameter was added, so the other
UA could not process the SIP message properly.
To fix this, the header parameter is ignored when parsing the
P-Preferred-Identity header.
This also applies to the P-Asserted-Identity header.

Bug: 427272461
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verified PUBLISH/INVITE request contains the
P-Preferred-Identity header without the header parameters with TISS)

Change-Id: Ib8ee7afac9a636541804cbfe53c8943d6a6bb6e7