(Aos) Retry Initial Registration on Same P-CSCF for USC
This change modifies the Re-Registration failure handling mechanism for the USC carrier.

To comply with the USC carrier requirement (IMS 1.2.4), the device must attempt an initial registration with the currently connected P-CSCF after a Re-Registration failure, before switching to a new P-CSCF.

This is implemented by adding the relevant error codes to the ims.rereg_retry_err_code_for_init_reg_with_same_pcscf_int_array asset for the USC carrier configuration. This will trigger a new initial registration attempt on the same P-CSCF for the specified failure types.

Requirement :
< IMS 1.2.4 >
The device must follow the procedures described in 3GPP TS 24.229 for dealing with failures during Re-Registration (5.1.1.4), and try the initial registration with the currently connected P-CSCF before initiating switch-over. The switch-over attempt to a different PCSCF is mandatory.

Bug: 438350123
Flag: NONE pre-production
Test: manual test using TISS simulator
Change-Id: I85ad231e1fd49eebda3097a7fb2188b65f5ac549