(core) Allow setting "+sip.instance" parameter for emergency session
In general, UE shall not include the "+sip.instance" header parameter in
the Contact header field in its non-REGISTER requests and responses
except when the request or response is guaranteed to be sent to a
trusted intermediary that will remove the "+sip.instance" header
parameter prior to forwarding the request or response to the
destination.
For emergency calls, the network may need to identify the IMEI to
perform the access transfer between PS and CS, so this change allows to
set the "+sip.instance" header parameter for an emergency session.
(Reference: 3GPP TS 24.237 and RFC 7255)

Bug: 444077944
Flag: EXEMPT BUGFIX
Test: manual (verify the emergency call test case for VDF-DE)
Change-Id: Ie93d8fcf3fb3eab220d9bbb95fbf741b61ee8ab3