(Mtc) Reduce call setup time by calling StartImsTraffic when the `IdleState` is created
In the previous revision, we tried to reduce the time taken by adjusting the order of the `IMtcBlockRule`s in the b/433142768.
Even with that, we were wondering if there was a way to make it even faster, as `StartImsTraffic` and its response takes the longest time among the `IMtcBlockRule`s.
As a result, will start this logic right after the `IdleState` is created.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulated with TISS(MO, MT, call swap, merge)
Bug: 433869994
Flag: NONE pre-production

Change-Id: Id80a9be137120f6fe36aee443456f6fd43f8a598