(isil) Preserve listener for callSessionTerminated before async call
When the IMS enabler is stopped and the MTC enabler is unbound during
outgoing call, the termination event is passed to the telephony through
ImsCallSessionCallback#invokeTerminated -> callSessionTerminated.

Right after that, ImsCallSessionCallback's listener is reset by
ImsCallSessionImpl#close.

But because invokeTerminated uses asynchronous call, the listener can be
reset before its callSessionTerminated is invoked. Consequently the
telephony cannot be notified call termination and the call stucks.

This CL preserves the listener before async call, to correctly invoke
callSessionTerminated even if mListener is reset.

Bug: 446774422
Flag: EXEMPT BUGFIX
Test: atest ImsStackJavaTests
Test: Manual test (see ag/36080024)
Change-Id: Ia2e526987bb0c6478d99234594793b88088f6eb2