(MTS) Notify native layer when MO SMS TR1 timer expires
This change introduces a mechanism to inform the native IMS stack when
the TR1 timer for a Mobile Originated (MO) SMS expires in the Java
SmsRLStateMachine.

Context: When an Emergency SMS (e-SMS) is sent, the SmsRLStateMachine
starts a T1 timer to await a response. If this timer expires, it
indicates a timeout in the SMS sending process. To ensure the Emergency
SMS state in AOS is correctly managed, the native layer
(MtsMessageController) needs to be informed to send a "done"
notification (NotifyEmergencySmsState(FALSE)) to AOS. This is essential
for the SCBM (SMS Callback Mode) handling.

Also handled Emergency SMS State in case of PDN disconnection

Test: atest ImsStackNativeTests ImsStackJavaTests
Test: Manual MO/MT SMS
Bug: 289441507
Doc: go/mts-notify-esms-aos
Flag: NONE pre-production
Change-Id: Ie7d3d1e29764ec46d5f685325830d53c32593231