(config) Update IMS configuration for SP40 Lab.
So far, the configurations for testing in the SP40 lab have been updated in the file corresponding to canonical_id 10039. However, we can not use it because canonical_id 10039 is now used for the "Orange FR NTN".
This CL updates the latest IMS configuration for SP40 Lab to the following file to be used instead.
carrier_config_mccmnc_262830.xml
carrier_config_mccmnc_262840.xml

Bug: 458564057
Test: build
Flag: NONE resource only update
Change-Id: Ie6825d166f2229076a8b39269eec9afd29bdb3cf