(its) Allow multiple headers with same name in ServerMessage
The ServerMessage.builder previously used a 'Map' to store headers, which
overwrote existing values when a header with the same name was added.

To comply with the SIP standard, which permits multiple headers with the
same name, the data structure has been changed to a 'List<Pair<String,
String>>'. This change ensures that all headers are preserved in the
order they are added, allowing for the creation of more
standard-compliant SIP messages.

Test: its MtCallTest using multiple header addition
Bug: 369467734
Flag: EXEMPT TEST_ONLY

Change-Id: If05c33d0ad1a94434f0bdb3b2bd6cc80239356db