(Isil) Remove delay for call end callbacks in the ISIL
The unnecessary delay of a call end callbacks in ISIL is being removed because it is no longer needed.
Currently, the Telephony framework and the IMS stack process using their own respective threads. Additionally, since it is now possible to determine if the Telephony framework is ready to respond to both outgoing and incoming calls, the delay is no longer necessary.

Bug: 453575422
Test: atest ImsStackJavaTests
Test: TISS simulation(MO, MT, call swap, merge, see the test list in the comment#13)
Flag: NONE pre-production
Change-Id: I4578b06837e6a0bc03e97ef5766bfc2b5b26e597