config: Add a Contact address for non-REGISTER messages for Orange
In the carrier lab test, incoming MESSAGE request (SMS over IP) doesn't
use the exact Contact address registered during IMS registration. This
results in the device sending a "404 Not Found" response.
To resolve this issue, this change is to allow the device to receive
incoming SIP requests by ensuring a distinct Contact address is
registered for all non-REGISTER messages, and set the user-info field in
the Contact address of non-REGISTER request to use the authorized AOR's
user-info value instead of UUID.

Bug: 464160052
Flag: EXEMPT BUGFIX
Test: manual (verify incoming SIP request with the Request-URI being set
to the authorized AOR's user-info value)

Change-Id: Iaaa091200885d0adaef46e96089a127b565df9a8