(Mtc) Allow precondition extension to USSI
This update enables support for the 'precondition' extension in USSI calls to ensure compatibility with specific carrier requirements.

According to standards, the 'precondition' extension should not be included in the Require header for USSI, as the precondition mechanism itself is unnecessary for USSI. Consequently, the MTC was previously configured to disable precondition support for all USSI calls, regardless of the overall device configuration.

However, it has been observed that certain carriers (e.g., AirTel IN) include the 'precondition' tag in the Require header of network-initiated USSI requests. This caused the MTC to reject these requests due to the unsupported extension.

To accommodate this behavior, the logic that explicitly disallowed the 'precondition' extension for USSI has been removed. The MTC will now accept USSI requests even if they contain the 'precondition' tag in the Require header.

Test: atest ImsStackNativeEnablerMtcTests
Test: TISS simulation for network initiated USSI
Bug: 468979445
Flag: NONE pre-production
Change-Id: I6da76274f258a8c06caaea81720633ac216633a6